Refs #36 - Move track list file.
