+++ application/src/track-list.js
@@ -1,0 +1,12 @@
+'use strict';
+
+var urlRoot = 'assets/tilemaps/maps/';
+
+module.exports = {
+    'Desert' : {
+        'Gravel and Pits'    : urlRoot + 'desert/gravel-and-pits.json',
+        'Square Loop'        : urlRoot + 'desert/square-loop.json',
+        'Figure 8'           : urlRoot + 'desert/figure8.json',
+        'Bathroom Obstacles' : urlRoot + 'desert/figure8wObs.json'
+    }
+};
